After checking out DF13 connections now that cables arrived, I discovered that TX/RX were reversed on the APM2 connection.  This is fixed in both the 1.0 and the 1.1 versions of the board.
On the radio modem connector I had grounded both pins 5 and 6 to accept
either a 5 or 6 position connector, but upon testing, if this is grounded
on a 6 position cable, it puts the radio modem into bootloader mode, so
I have removed the grounding connection on pin 5.  A 5 position connector
could still be used with a bit of extra work on the backside to ground pin 5.